FREESCAPE: moved NeoDecoder into the Freescape namespace
